convert url-shortener module to the new file/folder convention, rename features to modules
