+++ src/app/[slug]/page.tsx
@@ -5,7 +5,7 @@ import {
     RedirectType,
 } from 'next/navigation';
 
-import { getUrlEntry } from '@/features/url-shortener';
+import { getUrlEntry } from '@/modules/url-shortener';
 import { logger } from '@/lib/logger';
 
 import { UrlEntry } from '../../../generated/prisma';
