--- src/features/url-shortener/index.ts
@@ -1,3 +1,0 @@
-export * from './create-short-url-form';
-export { getUrlEntry } from './functions';
-export * from './types';
