--- src/features/url-shortener/types.ts
@@ -1,5 +1,0 @@
-export type Slug = string;
-
-export interface ShortenUrlResponse {
-    url: string;
-}
