+++ src/modules/register/definitions.ts
