+++ src/modules/url-shortener/create-short-url-form.tsx
@@ -2,7 +2,6 @@
 
 import { useState } from 'react';
 
-import { z } from 'zod';
 import { toast } from 'sonner';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
@@ -24,38 +23,31 @@ import {
 } from '@/components/ui/form';
 
 import { performRequest } from '@/lib/api';
-
-const formSchema = z.object({
-    url: z.url({
-        protocol: /^https?$/,
-        hostname: z.regexes.domain,
-    }),
-    slug: z.string().max(32),
-});
-
-type ShortUrlFormData = z.infer<typeof formSchema>;
-
-type ShortUrlResponse = { message: string; url: string };
+import {
+    shortenUrlSchema,
+    type ShortenUrlFormData,
+    type ShortenUrlSuccess,
+} from './definitions';
 
 export function CreateShortUrlForm() {
     const [loading, setLoading] = useState(false);
     const [shortenedUrl, setShortenedUrl] = useState<string | undefined>(
         undefined
     );
-    const form = useForm<ShortUrlFormData>({
-        resolver: zodResolver(formSchema),
+    const form = useForm<ShortenUrlFormData>({
+        resolver: zodResolver(shortenUrlSchema),
         defaultValues: {
             url: '',
             slug: '',
         },
     });
 
-    const onSubmit = async (formData: ShortUrlFormData) => {
+    const onSubmit = async (formData: ShortenUrlFormData) => {
         setLoading(true);
 
         const response = await performRequest<
-            ShortUrlFormData,
-            ShortUrlResponse
+            ShortenUrlFormData,
+            ShortenUrlSuccess
         >('/api/url-entry', 'POST', formData);
 
         if (response.status === 'success') {
@@ -102,6 +94,7 @@ export function CreateShortUrlForm() {
                             </FormItem>
                         )}
                     />
+
                     <FormField
                         control={form.control}
                         name='slug'
@@ -118,6 +111,7 @@ export function CreateShortUrlForm() {
                             </FormItem>
                         )}
                     />
+
                     <Button
                         className='w-1/2 cursor-pointer'
                         variant='default'
